# api/model_car.go
/*
 * Car info
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 0.0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package api

type Car struct {
	RegNum string `json:"regNum"`
	Mark string `json:"mark"`
	Model string `json:"model"`
	Year int32 `json:"year,omitempty"`
	Owner People `json:"owner"`
}